chore: move stories next to code.
