+++ react-app/src/component/Intro.mdx
@@ -14,11 +14,11 @@ Legends are visible representations of the mapping between colors and data value
 
 ### Continuous Legend component
 
-The continuous legend consists of a color ramp and a numeric scale indicating color values. The main props for this legend are value range(min and max), colortables(contains point and color) and few other optional paramater.
+The continuous legend consists of a color ramp and a numeric scale indicating color values. The main props for this legend are value range(min and max), colortables(contains point and color) and few other optional parameter.
 
 ### Discrete Legend component
 
-The discrete legends display a discrete list of elements in the given hierarchy. The main props for this legend are data(discreteData), colortables(contains point and color) and few other optional paramater.
+The discrete legends display a discrete list of elements in the given hierarchy. The main props for this legend are data(discreteData), colortables(contains point and color) and few other optional parameter.
 
 ### Color Selector Component
