+++ react-app/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "rootDir": ".",
+    "rootDir": "src",
     "outDir": "dist",
     "module": "es2022",
     "target": "es2022",
